% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.r
\name{readwrite}
\alias{readwrite}
\title{Import/Export any file}
\usage{
readwrite(filepath, writepath, varname = TRUE, sheet = 1, skip = 0, write = FALSE, ...)
}
\arguments{
\item{filepath}{The path of the file that is to be imported. In case of exporting a file, it is the file that is to be exported.}

\item{writepath}{Only when exporting (write = \code{TRUE}). The path where the file is to be exported.}

\item{varname}{Default set to \code{TRUE}, suggesting that first row contains variable names. Set to \code{FALSE} if there is no variable name in the first row.}

\item{sheet}{In case of importing an excel file, it refers to the sheet index that is to be imported from the file. In case of exporting an excel file, it refers to the name of the sheet (in double quotes) that you would like to give.}

\item{skip}{In case of importing an excel file, it refers to the number of rows to be skipped while importing. Starts from 1 which is the second row, as first row is considered as column name, if \code{varname} is set \code{TRUE}.}

\item{write}{If importing, set it to \code{FALSE} (default). If exporting, set it to \code{TRUE}.}

\item{...}{Arguments passed on to \code{xlsx::read.xlsx}
\describe{
  \item{sheetName}{a character string with the sheet name.}
  \item{rowIndex}{a numeric vector indicating the rows you want to
    extract.  If \code{NULL}, all rows found will be extracted, unless
  \code{startRow} or \code{endRow} are specified.}
  \item{colIndex}{a numeric vector indicating the cols you want to
    extract.  If \code{NULL}, all columns found will be extracted.}
  \item{as.data.frame}{a logical value indicating if the result should
    be coerced into a \code{data.frame}.  If \code{FALSE}, the result is
    a list with one element for each column.}
  \item{colClasses}{For \code{read.xlsx} a character vector that
	represent the class of each column.  Recycled as necessary, or if
	the character vector is named, unspecified values are taken to be
	\code{NA}.  For \code{read.xlsx2} see \code{\link{readColumns}}.}
  \item{keepFormulas}{a logical value indicating if Excel formulas
    should be shown as text in \R and not evaluated before bringing them
    in.}
  \item{encoding}{encoding to be assumed for input strings.  See
	\code{\link[utils]{read.table}}.}
  \item{startRow}{a number specifying the index of starting row.  For
    \code{read.xlsx} this argument is active only if
    \code{rowIndex} is \code{NULL}.}
  \item{endRow}{a number specifying the index of the last row to pull.  If
	\code{NULL}, read all the rows in the sheet.  For
    \code{read.xlsx} this argument is active only if
    \code{rowIndex} is \code{NULL}.}
}}
}
\description{
Imports CSV, XLS, XLSX, DTA (stata), and SAS7BDAT (sas) files.
Exports to CSV, XLS, XLSX files.
It auto detects the file extension in the \code{filepath} or \code{writepath} to read/write the file in your desired extension.
}
\details{
This functions reads CSV, XLS, XSLX, DTA (stata files), SAS7BDAT (sas files), and writes CSV, XLS, XLSX files. Can be used to Import and Export data frames.
}
\note{
Only data frames (\code{data.frame}) can be exported. This function coerces the table in a data frame, while writing it to a file.

The fuction relies on Java (with the same architechture as your R) to export files. Credit goes to the authors of "xlsx" package.
}
\examples{
\donttest{
##Import
file <- readwrite("C:/Users/Personal/Documents/file.xlsx", varname = TRUE,
    sheet = 2)
## Export
readwrite(file, "C:/Users/Personal/Documents/filesheet2.xlsx",
    sheet = "sheet name", write = TRUE)
}
}
