% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.R
\name{readwrite}
\alias{readwrite}
\title{Import/Export any file}
\usage{
\code{readwrite(filepath, writepath, varname = TRUE, sheet = 1, skip = 0,
    read = TRUE, write = FALSE)}
}
\arguments{
\item{filepath}{The path of the file that is to be imported. In case of exporting a file, it is the file that is to be exported.}

\item{writepath}{Only when exporting (write = \code{TRUE}). The path where the file is to be exported.}

\item{varname}{Default set to \code{TRUE}, suggesting that first row contains variable names. Set to \code{FALSE} if there is no variable name in the first row.}

\item{sheet}{In case of importing an excel file, it refers to the sheet that is to be imported from the file. In case of exporting an excel file, it refers to the name of the sheet (in double quotes) that you would like to give.}

\item{skip}{In case of importing an excel file, it refers to the number of rows to be skipped while importing. Starts from 1 which is the second row, as first row is considered as column name, if \code{varname} is set \code{TRUE}.}

\item{read}{If importing, set it to \code{TRUE} (default). If exporting, set it to \code{FALSE}.}

\item{write}{If importing, set it to \code{FALSE} (default). If exporting, set it to \code{TRUE}.}
}
\description{
Imports CSV, XLS, XLSX, DTA (stata), and SAS7BDAT (sas) files.
Exports to CSV, XLS, XLSX files.
It auto detects the file extension in the \code{filepath} or \code{writepath} to read/write the file in your desired extension.
}
\details{
This functions reads CSV, XLS, XSLX, DTA (stata files), SAS7BDAT (sas files), and writes CSV, XLS, XLSX files. Can be used to Import and Export data frames.
}
\note{
Only data frames (\code{data.frame}) can be exported. This function coerces the table in a data frame, while writing it to a file.

The fuction relies on Java (with the same architechture as your R) to export files. Credit goes to the authors of "xlsx" package.
}
\examples{
##Import
file <- readwrite("C:/Users/Personal/Documents/file.xlsx", varname = TRUE,
    sheet = 2)
## Export
readwrite(file, "C:/Users/Personal/Documents/filesheet2.xlsx",
    sheet = "sheet name", read = FALSE, write = TRUE)
}
